♻️ [refactor] : Chat Server Config 파일명 변경 및 모듈화
