+++ chat_server/config/socket.config.ts
@@ -1,0 +1,10 @@
+import { ServerOptions } from "socket.io";
+
+const options: Partial<ServerOptions> = {
+	cors: {
+		origin: process.env.SERVER_ADDRESS,
+		credentials: true,
+	},
+};
+
+export default { options };
